backend/pyproject.toml
[project]
name = "robo-data-fabric"
version = "0.1.0"
description = "Backend API for Data Fabric - DataSource management and MindsDB integration"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "requests>=2.31.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "neo4j>=5.0.0",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
]
